+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - [2650](https://github.com/influxdb/influxdb/pull/2650): Add SHOW GRANTS FOR USER statement. Thanks @n1tr0g.
 - [3013](https://github.com/influxdb/influxdb/issues/3013): Panic error with inserting values with commas
+- [3125](https://github.com/influxdb/influxdb/pull/3125): Graphite Input Protocol Parsing
 
 ### Bugfixes
 
@@ -28,6 +29,8 @@
 - [#2944](https://github.com/influxdb/influxdb/issues/2944): Don't require "WHERE time" when creating continuous queries.
 - [#3075](https://github.com/influxdb/influxdb/pull/3075): GROUP BY correctly when different tags have same value.
 - [#3078](https://github.com/influxdb/influxdb/pull/3078): Fix CLI panic on malformed INSERT.
+- [2102](https://github.com/influxdb/influxdb/issues/2102): Re-work Graphite input and metric processing
+- [2996](https://github.com/influxdb/influxdb/issues/2996): Graphite Input Parsing
 
 ## v0.9.0 [2015-06-11]